make go-get-able (commit w/ patch applied)
To enable "go get" I committed the printer/ code with the patch applied.

The patch is kept to use when updating the printer code from the latest
go sources. To update the patch we keep around a set of pristine source
files (in printer/orig/) and generate new patches using it.

The Makefile has been updated with targets for this new flow.